Add support for custom dictionary  types in cached entities (#145)
* Add support for custom dictionary  types in cached entities

* Code cleanup

* Use ToDictionary instead of ToImmutableDictionary

* Improvements

* Rename `GatewayClientCache` back

* Enhance type safety in Cast method of GatewayClientCache

Modified the `Cast<TKey, TElement>` method to include a type check for `ImmutableDictionary<TKey, TElement>`. The method now verifies if the input `dictionary` is of the correct type before casting. If the check fails, an `ArgumentException` is thrown, improving error handling and type safety.

* Refactor generic type parameters in collections

Updated the `GatewayClientCache` and `CollectionsUtils` classes to replace the generic type parameter `TElement` with `TValue` in the `CreateDictionary` and `CreateImmutableDictionary` methods. This change enhances naming consistency and clarifies the intended use of these methods, while maintaining the existing type constraints for `TKey` and `TValue`.

* Improve error message in Cast method

Updated the exception message in the `Cast` method to provide clearer guidance on the requirement for the dictionary to be created using the `CreateDictionary<TSource, TKey, TValue>` method. This change enhances clarity for users encountering this exception.

* Cleanup

* Refactor Guild and RestGuild for JSON model updates

- Introduced `UpdateJsonModel` in `Guild` to update
  properties of `JsonGuild` based on the current state.
- Added `Clone` method in `JsonGuild` for shallow copies.
- Updated `RestGuild` to use `JsonGuild` type and modified
  `JsonModel` property to return a cloned instance.
- Adjusted constructor of `RestGuild` to accept `JsonGuild`.
- Made `UpdateJsonModel` in `RestGuild` virtual for
  potential overrides in derived classes.

* Refactor exception handling in GatewayClientCache

- Added using directive for System.Diagnostics.CodeAnalysis.
- Replaced direct ArgumentException with a new method
  ThrowInvalidDictionary() that throws InvalidOperationException.
- Improved code organization and maintainability by encapsulating
  exception logic.
- Modified Cast method to return null after calling
  ThrowInvalidDictionary(), although this return may not be reached.

* Refactor Guild class for improved JSON model handling

The `Guild` class now implements the `ICloneable` interface more effectively by replacing the `UpdateJsonModel` method with a new `GetJsonModel` method that returns a cloned `JsonGuild` object. This change streamlines the process of obtaining the JSON model representation of the guild. The `RestGuild` class has also been updated to utilize the new method, ensuring consistency across the implementation.