Implement minimal API builders (#159)
* Improve `MethodsForPropertiesGenerator`

* Fix the source generator

* Add builders for application commands

* Add text command builders

* Add component interaction builders

* Fix build and add a missing GenerateMethodsForPropertiesAttribute

* Improvements

* Remove commented code

* Remove unused NC0001 suppressions

* Add an `IApplicationCommandsBuilder<...>` with an autocomplete context

* Remove obsolete members

* Rename DefaultGuildUserPermissions to DefaultGuildPermissions

This commit updates the NetCord library to rename the
`DefaultGuildUserPermissions` property to `DefaultGuildPermissions`
across various classes and files. This change standardizes the
definition of permissions for application commands. The updates
include modifications to attributes, command builders, command
info classes, and command properties, while preserving the
context for commands and maintaining the overall structure of
the command-related classes.