Add xml docs for NetCord.Services (#164)
* Add xml docs for builders and partial docs for contexts

* Add more xml docs for contexts

* Add (I)ApplicationCommandService xml docs

* Add xml docs for IApplicationCommandInfo and small refactor

* Resolve merge conflict resolve errors

* Fix whitespace

* Clarify AddModules method documentation

Updated the XML documentation for the `AddModules` method in the `IService` interface to specify that it scans for **public modules** in the provided assembly. This change improves clarity for developers without altering the method's functionality or behavior.

* Add xml docs for `ApplicationCommandServiceManager`

* Add XML documentation to IApplicationCommandService

Added a summary comment to the `IApplicationCommandService` interface, describing it as the base interface for `ApplicationCommandService<TContext>`. This improves code readability and provides better context for developers using documentation tools like IntelliSense.

* Add xml docs to app command modules

* Add xml docs to component interaction modules

* Improve XML documentation and enhance code clarity

Updated XML documentation across `ApplicationCommandService.cs`,
`ComponentInteractionService.cs`, and `IComponentInteractionService.cs`
to improve clarity and maintainability.

- Used `<inheritdoc>` for parameter documentation in `ApplicationCommandService`.
- Added `<summary>`, `<typeparam>`, `<param>`, and `<returns>` tags
  to describe class purposes, methods, and properties in
  `ComponentInteractionService` and `IComponentInteractionService`.
- Enhanced developer experience with detailed descriptions of
  functionality and parameters.

* Add xml docs for command service

* Add xml docs for command modules

* Make base modules abstract

* Add xml docs for UserId

* Add xml docs for command and component interaction builders

* Remove `abstract` modifier from `IApplicationCommandInfo.Type`

* Remove xml docs from `IApplicationCommandInfo`

* Reformat `ICommandInfo`

* Simplify XML documentation for parameters

Updated XML documentation in `CommandBuilder.cs` and
`ComponentInteractionBuilder.cs` to simplify `<inheritdoc>`
tags for parameters. Replaced or adjusted references to
summary paths for `aliases` and `customId` parameters to
improve clarity and maintain consistency.